Rag/first_rag/requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-huggingface==0.0.3
langchain-groq==0.1.9
streamlit==1.39.0
python-dotenv==1.0.1
chardet==5.2.0
sentence-transformers==3.1.1
pypdf==5.0.1